Parse into fields, not actual date/time values
This completely separates the parsing module from the local-datetimes module. Parsing now returns a set of fields, which can be put together by a datetime's constructor.

As an (intended) side-effect, it more correctly shows where an error came from: whether it was during parsing, or trying to read an integer, or that one of the fields was out of range (such as day 33 of a month)

Some of the tests were casualties of this. I will add them back in soon.