Store only the total offset in a transition
This is being changed to allow compiled time zones, which only store the total offset, to use the same structs as the zoneinfo-data crate.

A workaround would be to only use one of the fields for the total offset in that case, and leave the other one blank -- the effect is the same, since only the `total_offset` method is actually called, rather than the fields -- but they would be nominally incorrect.