convert_local impl for ZonedDateTime
This was one of the last ‘big missing pieces’ of the time zone puzzle. It allows users to say “the clock on the wall says 5pm; what time is it in UTC?”—basically doing the *reverse* operation to the usual one when you want to display a time.

It’s trickier than it sounds, though, as the operation can fail as the resulting times may be ambiguous or impossible, which is why it requires *so much* code.