Make TimeZone a struct, not a trait
The TimeZone trait is now just a struct, as it was discovered that there’ll only be one actual type of time zone, at least for now: one with a name and a number of transitions. Both the zoneinfo-data and zoneinfo-compiled crates will use this struct.

This means that all the impl code (rather than imports) in the zoned module root can be moved to a new zoned submodule, because the entirety of it pertains to that struct!

ZoneSet has been renamed to FixedTimespanSet, and ZoneDetails is now a FixedTimespan. This makes more sense:

- How can one ZoneSet have more than one ZoneDetails? The name implied that it would be the details of the entire zone, rather than the details of *part* of the zone.
- The terminology of what constitutes a time zone and what constitutes a period with a fixed offset is now more clear.
- It also shows that one rule can only apply for a certain amount of time (a FixedTimespan, in fact).