Make Offset a struct
Having it be a struct instead of an enum means that we can encapsulate the number of seconds, which an enum would automatically make public. This means that offsets will always be constructed to be within their allowed span of seconds.

Also, remove a few outdated comments here and there, which stem from the era when I didn’t know the difference between an offset and a time zone.