Move everything system-related into its own module
This fuses the ‘now’ module with the ‘cal::zonefinder’, which may seem completely unrelated, but:

- It now means that everything that requires system calls or poking around in various system files is now in its own module, which both the ‘get current time’ function and the ‘get current timezone’ function are;
- The ‘get current timezone’ function, believe it or not, actually has nothing to do with time zones—it only returns the *name* of the time zone, and has nothing to do with dates or months or years, so it doesn’t belong in the cal module with all the other timezone functions.

I’ve decided to leave the ‘sys_’ prefix in, despite being in the system module, as it seems more *systemy* that way. That’s no joke. It implies that it’s an internal function that shouldn’t really be user-facing. Liner notes! Maybe this will come back to bite me some day.