Adjust zone to allow for runtime TimeZones
The old implementation was hampered in that all the time zones had to be borrowed, rather than owned. Unfortunately, due to Rust’s lifetime rules, it’s not possible to store a FixedTimezoneSet within an OwnedFixedTimezoneSet, because it holds a vector of strings which you can’t convert to a slice of string slices.

So instead, each time zone now knows where it came from: either statically (or with a lifetime), or created at runtime. This simplifies the time-zone code, as TimeZone now no longer needs a lifetime (it’ll work with static data or data contained within an Arc, neither of which need a stack frame), but more importantly, it allows time zones to be created at runtime (such as from a file) and used just like any other time zone.